Make test_misc compile with -Werror again.
Commit 5978a46e3 (master)/3402f9aa0 (release/2.7) removed the
local prototype/function get_random in test_misc and use the
global prototype in crypto.h.

This made get_random return uint64_t instead of uint32_t.
Adjust the calls in test_list accordingly.

On master the general hash rewrite/cleanup completely
removes the iv argument from hash_init and therefore
only 2.7 requires this patch.

Acked-by: Gert Doering <gert@greenie.muc.de>
Gerrit URL: https://gerrit.openvpn.net/c/openvpn/+/1839
Message-Id: <20260806111323.32150-1-gert@greenie.muc.de>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg38167.html

+++ tests/unit_tests/openvpn/test_misc.c
@@ -170,8 +170,8 @@ test_list(void **state)
      */
 
     struct gc_arena gc = gc_new();
-    struct hash *hash = hash_init(10000, get_random(), word_hash_function, word_compare_function);
-    struct hash *nhash = hash_init(256, get_random(), word_hash_function, word_compare_function);
+    struct hash *hash = hash_init(10000, (uint32_t)get_random(), word_hash_function, word_compare_function);
+    struct hash *nhash = hash_init(256, (uint32_t)get_random(), word_hash_function, word_compare_function);
 
     printf("hash_init n_buckets=%u mask=0x%08x\n", hash->n_buckets, hash->mask);
